pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "evanesco-offline"
version = "0.1.0"
description = "Fully local/offline document anonymization (OCR + spaCy/regex + optional Ollama GPT-OSS)"
authors = [{name="You"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=10.0.0",
    "pdf2image>=1.17.0",
    "pytesseract>=0.3.10",
    "spacy>=3.7.2",
    "regex>=2024.5.15",
    "click>=8.1.7",
    "tqdm>=4.66.4",
    "python-dotenv>=1.0.1",
    "pydantic>=2.8.0",
    "rich>=13.7.1",
    "reportlab>=4.1.0",
    "img2pdf>=0.5.1",
    "numpy>=1.26.0",
    "opencv-python-headless>=4.9.0.80",
    "rapidfuzz>=3.9.0",
    "pymupdf>=1.24.8",
    "typer>=0.12.3",
    "orjson>=3.10.7",
    "requests>=2.32.3",
    "pandas>=2.3.2",
    "gradio>=4.44.0",
    "PyYAML>=6.0.1",
]

[project.scripts]
evanesco = "evanesco.cli:app"
evanesco-ui = "evanesco.ui:launch"
evanesco-api = "evanesco.api:run"

[project.optional-dependencies]
docs = [
    "sphinx>=7.0.0",
    "myst-parser>=2.0.0",
    "sphinx-autodoc-typehints>=2.2.1",
    "sphinx-rtd-theme>=2.0.0",
]
server = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "prometheus-client>=0.20.0",
]
nlp = [
    "langdetect>=1.0.9",
]
